requirements.txt
Django==2.0
django-cors-middleware==1.3.1
gitdb2==2.0.3
GitPython==2.1.8
mysqlclient==1.3.12
psycopg2==2.7.3.2
pytz==2017.3
smmap2==2.0.3
